[BOOST-715] rename files to kebab-case
