+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/dapr-manager.ts
@@ -2,8 +2,8 @@ import * as path from 'path'
 import * as fs from 'fs'
 import * as util from 'util'
 import { stateStore } from './templates/statestore'
-import { K8sManagement } from './k8s-sdk/K8sManagement'
-import { DaprTemplateValues } from './templates/templateInterface'
+import { K8sManagement } from './k8s-sdk/k8s-management'
+import { DaprTemplateValues } from './templates/template-types'
 import { HelmManager } from './helm-manager'
 import * as Mustache from 'mustache'
 import { BoosterConfig } from '@boostercloud/framework-types'
