+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/deploy-manager.ts
@@ -1,12 +1,12 @@
-import { K8sManagement } from './k8s-sdk/K8sManagement'
+import { K8sManagement } from './k8s-sdk/k8s-management'
 import { BoosterConfig } from '@boostercloud/framework-types'
 import { getProjectNamespaceName, createProjectZipFile, createIndexFile, uploadFile } from './utils'
-import { uploadService } from './templates/uploadService'
-import { boosterVolumeClaim } from './templates/boosterVolumeClaim'
-import { boosterService } from './templates/boosterService'
-import { Template, TemplateValues } from './templates/templateInterface'
-import { uploaderPod } from './templates/fileUploader'
-import { boosterAppPod } from './templates/boosterApp'
+import { uploadService } from './templates/upload-service-template'
+import { boosterVolumeClaim } from './templates/volume-claim-template'
+import { boosterService } from './templates/booster-service-template'
+import { Template, TemplateValues } from './templates/template-types'
+import { uploaderPod } from './templates/file-uploader-app-template'
+import { boosterAppPod } from './templates/booster-app-template'
 import { HelmManager } from './helm-manager'
 import { DaprManager } from './dapr-manager'
