+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/booster-app-template.ts
