+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/booster-service-template.ts
