+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/file-uploader-app-template.ts
