+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/index-template.ts
