+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/template-types.ts
