+++ packages/framework-provider-kubernetes-infrastructure/src/infrastructure/templates/upload-service-template.ts
